code/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
fastapi==0.75.0
Flask==2.0.3
gTTS==2.2.3
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
Pillow==9.0.1
playsound==1.3.0
pydantic==1.9.0
pyparsing==3.0.7
pytesseract==0.3.9
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
watchgod==0.8.1
websockets==10.2
Werkzeug==2.0.3
